# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "avatar-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/panel/avatars": {
      "get": {
        "summary": "List Avatars",
        "operationId": "list_avatars_panel_avatars_get",
        "parameters": [
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvatarOut"
                  },
                  "title": "Response List Avatars Panel Avatars Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Avatar",
        "operationId": "create_avatar_panel_avatars_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_avatar_panel_avatars_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/panel/avatars/{avatar_id}": {
      "get": {
        "summary": "Get Avatar",
        "operationId": "get_avatar_panel_avatars__avatar_id__get",
        "parameters": [
          {
            "name": "avatar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Avatar Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/worker-status": {
      "get": {
        "summary": "Worker Status",
        "operationId": "worker_status_api_v1_worker_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerStatusOut"
                }
              }
            }
          },
          "401": {
            "description": "Chave de API ausente ou inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado ou de outra chave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de intenção ou vídeo indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos ou fala acima do limite.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Banco ou receita vigente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs": {
      "post": {
        "summary": "Create Job",
        "operationId": "create_job_api_v1_jobs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewVideoJob"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreatedOut"
                }
              }
            }
          },
          "401": {
            "description": "Chave de API ausente ou inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado ou de outra chave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de intenção ou vídeo indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos ou fala acima do limite.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Banco ou receita vigente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "401": {
            "description": "Chave de API ausente ou inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado ou de outra chave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de intenção ou vídeo indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos ou fala acima do limite.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Banco ou receita vigente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs/{job_id}/download": {
      "get": {
        "summary": "Download Job",
        "operationId": "download_job_api_v1_jobs__job_id__download_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "video/mp4": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Chave de API ausente ou inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado ou de outra chave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de intenção ou vídeo indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos ou fala acima do limite.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Banco ou receita vigente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "206": {
            "description": "Intervalo solicitado por Range.",
            "content": {
              "video/mp4": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "416": {
            "description": "Intervalo Range fora do arquivo."
          }
        },
        "security": [
          {
            "ApiKeyBearer": []
          }
        ]
      }
    },
    "/api/v1/avatars": {
      "get": {
        "summary": "List Avatars",
        "operationId": "list_avatars_api_v1_avatars_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarListOut"
                }
              }
            }
          },
          "401": {
            "description": "Chave de API ausente ou inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado ou de outra chave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de intenção ou vídeo indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos ou fala acima do limite.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Banco ou receita vigente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyBearer": []
          }
        ]
      }
    },
    "/api/v1/scenes": {
      "get": {
        "summary": "List Scenes",
        "operationId": "list_scenes_api_v1_scenes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneListOut"
                }
              }
            }
          },
          "401": {
            "description": "Chave de API ausente ou inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado ou de outra chave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de intenção ou vídeo indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos ou fala acima do limite.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Banco ou receita vigente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyBearer": []
          }
        ]
      }
    },
    "/api/v1/avatars/{avatar_id}/preview": {
      "get": {
        "summary": "Avatar Preview",
        "operationId": "avatar_preview_api_v1_avatars__avatar_id__preview_get",
        "parameters": [
          {
            "name": "avatar_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Avatar Id"
            }
          },
          {
            "name": "aspect_ratio",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "9:16",
                "16:9"
              ],
              "type": "string",
              "default": "9:16",
              "title": "Aspect Ratio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Chave de API ausente ou inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado ou de outra chave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de intenção ou vídeo indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos ou fala acima do limite.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Banco ou receita vigente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyBearer": []
          }
        ]
      }
    },
    "/api/v1/scenes/{scene_id}/preview": {
      "get": {
        "summary": "Scene Preview",
        "operationId": "scene_preview_api_v1_scenes__scene_id__preview_get",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          },
          {
            "name": "aspect_ratio",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "9:16",
                "16:9"
              ],
              "type": "string",
              "default": "9:16",
              "title": "Aspect Ratio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Chave de API ausente ou inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado ou de outra chave.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de intenção ou vídeo indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos ou fala acima do limite.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Banco ou receita vigente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvatarListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicAvatarOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AvatarListOut"
      },
      "AvatarOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "voice": {
            "type": "string",
            "title": "Voice"
          },
          "status": {
            "type": "string",
            "enum": [
              "preparando",
              "ativo",
              "falha",
              "arquivado"
            ],
            "title": "Status"
          },
          "prepare_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prepare Error"
          },
          "authorized_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorized At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "voice",
          "status",
          "prepare_error",
          "authorized_at",
          "created_at",
          "archived_at"
        ],
        "title": "AvatarOut"
      },
      "Body_create_avatar_panel_avatars_post": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          },
          "authorization_confirmed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorization Confirmed"
          }
        },
        "type": "object",
        "required": [
          "name",
          "file"
        ],
        "title": "Body_create_avatar_panel_avatars_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/PublicErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "PublicErrorOut"
      },
...
      "WorkerStatusOut": {
        "properties": {
          "last_heartbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Heartbeat At"
          }
        },
        "type": "object",
        "required": [
          "last_heartbeat_at"
        ],
        "title": "WorkerStatusOut"
      }
    },
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Chave criada no painel."
      }
    }
  }
}